commit c2cbc6b847acc8ffa2560958fbd72b70f21e5afe
    16705 Want lifetime extensions for TCP_MD5SIG SAs
commit afa937a6c8732de9a6e3c2830b45f93829ff5aff
    16709 smbios tests should also be built 64-bit
commit 49f72993927fbcc8093ddd143715f814931de21a
    16708 SMBIOS 3.8 Support
commit 30fdefb82f3ce27d19928cb02b9a265b2c321a49
    16693 dumpadm: want parsable size estimate
commit 4763305e3243687c189d755d737d52205b2614ed
    16675 want syncfs(3C)
commit ea7b7d8ad0ddfa6a32fb675e11cc106f686a5bb5
    16676 sockfs should not use the rootvfs vfs_t
commit 1decd07c6adb97a56572371aa1bb958cec8a89df
    16704 protolist corrupts inode numbers larger than 2^31
commit 5e302d10a5edb261c2788efeac96e18a97a33d08
    16703 types.h.3head has incorrect type for LP64 ino_t
commit d3de9ea6ceda8b5560cfee4dc7f4064e1de93c19
    16697 krb5: dangling pointer 'stash_file' to 'stashbuf' may be used

Conflicts:
    manifest
    usr/src/cmd/dumpadm/Makefile
    usr/src/cmd/dumpadm/dconf.h
    usr/src/cmd/dumpadm/main.c
    usr/src/man/man8/dumpadm.8
    usr/src/uts/common/fs/tmpfs/tmp_vfsops.c

+++ usr/src/uts/common/sys/vfs_opreg.h
@@ -54,8 +54,8 @@ typedef union fs_func {
  * File systems use arrays of fs_operation_def structures to form
  * name/value pairs of operations.  These arrays get passed to:
  *
- * 	- vn_make_ops() to create vnodeops
- * 	- vfs_makefsops()/vfs_setfsops() to create vfsops.
+ *	- vn_make_ops() to create vnodeops
+ *	- vfs_makefsops()/vfs_setfsops() to create vfsops.
  */
 typedef struct fs_operation_def {
 	char *name;			/* name of operation (NULL at end) */
@@ -76,9 +76,9 @@ typedef struct fs_operation_def {
  */
 typedef struct fs_operation_trans_def {
 	char *name;			/* name of operation (NULL at end) */
-	int offset;			/* byte offset within ops vector */
+	size_t offset;			/* byte offset within ops vector */
 	fs_generic_func_p defaultFunc;	/* default function */
-	fs_generic_func_p errorFunc; 	/* error function */
+	fs_generic_func_p errorFunc;	/* error function */
 } fs_operation_trans_def_t;
 
 /*
